services/customer/requirements.txt
grpcio>=1.56.0
grpcio-tools>=1.56.0